# Addressing memory leak with SetDLLDirectory.