--- a/benchmarks/results/thumbnail.net48/FileOnQ.Imaging.Heif.Benchmarks.Thumbnail-report.html
+++ b/benchmarks/results/thumbnail.net48/FileOnQ.Imaging.Heif.Benchmarks.Thumbnail-report.html
@@ -2,7 +2,7 @@
 <html lang='en'>
 <head>
 <meta charset='utf-8' />
-<title>FileOnQ.Imaging.Heif.Benchmarks.Thumbnail-20220112-211740</title>
+<title>FileOnQ.Imaging.Heif.Benchmarks.Thumbnail-20220112-235445</title>
 
 <style type="text/css">
 	table { border-collapse: collapse; display: block; width: 100%; overflow: auto; }
@@ -23,12 +23,12 @@ UnrollFactor=1
 </code></pre>
 
 <table>
-<thead><tr><th>      Method</th><th>Mean</th><th>Error</th><th>StdDev</th><th>Allocated native memory</th><th>Native memory leak</th><th>Allocated</th>
+<thead><tr><th>      Method</th><th>Mean</th><th>Error</th><th>StdDev</th><th>Median</th><th>Allocated native memory</th><th>Native memory leak</th><th>Allocated</th>
 </tr>
-</thead><tbody><tr><td>Thumbnail_Write</td><td>60.80 ms</td><td>1.188 ms</td><td>2.658 ms</td><td>5,124,251 B</td><td>180 B</td><td>74,504 B</td>
-</tr><tr><td>Thumbnail_ToArray</td><td>60.41 ms</td><td>1.086 ms</td><td>1.592 ms</td><td>5,123,923 B</td><td>180 B</td><td>74,504 B</td>
-</tr><tr><td>Thumbnail_ToSpan</td><td>60.79 ms</td><td>1.181 ms</td><td>1.655 ms</td><td>5,123,923 B</td><td>180 B</td><td>-</td>
-</tr><tr><td>Thumbnail_ToStream</td><td>58.00 ms</td><td>1.130 ms</td><td>1.429 ms</td><td>5,123,923 B</td><td>180 B</td><td>140,816 B</td>
+</thead><tbody><tr><td>Thumbnail_Write</td><td>59.36 ms</td><td>1.167 ms</td><td>1.091 ms</td><td>59.51 ms</td><td>5,123,891 B</td><td>-</td><td>74,504 B</td>
+</tr><tr><td>Thumbnail_ToArray</td><td>59.17 ms</td><td>1.157 ms</td><td>2.364 ms</td><td>58.64 ms</td><td>5,123,307 B</td><td>-</td><td>74,504 B</td>
+</tr><tr><td>Thumbnail_ToSpan</td><td>59.17 ms</td><td>1.047 ms</td><td>1.779 ms</td><td>58.91 ms</td><td>5,123,579 B</td><td>-</td><td>-</td>
+</tr><tr><td>Thumbnail_ToStream</td><td>60.52 ms</td><td>1.209 ms</td><td>2.851 ms</td><td>59.45 ms</td><td>5,123,563 B</td><td>-</td><td>140,816 B</td>
 </tr></tbody></table>
 </body>
 </html>