--- a/src/FileOnQ.Imaging.Heif/Interop/Interop.cs
+++ b/src/FileOnQ.Imaging.Heif/Interop/Interop.cs
@@ -1,5 +1,6 @@
 ﻿#if NET48_OR_GREATER
 using System;
+using System.IO;
 using System.Runtime.InteropServices;
 
 namespace FileOnQ.Imaging.Heif
@@ -9,14 +10,19 @@ namespace FileOnQ.Imaging.Heif
 		[DllImport("kernel32.dll")]
 		private static extern bool SetDllDirectory(string dllToLoad);
 
+		private static bool loaded;
+
 		internal static void UpdateDllSearchPath()
 		{
-			var binPath = new Uri(typeof(LibHeif).Assembly.CodeBase).LocalPath;
-			var folder = System.IO.Path.GetDirectoryName(binPath);
-			var runtime = $"runtimes\\win-{RuntimeInformation.ProcessArchitecture}\\native";
-			var nativePath = System.IO.Path.Combine(folder, runtime);
+			if (loaded)
+				return;
 
-			var loaded = SetDllDirectory(nativePath);
+			var binPath = new Uri(typeof(LibHeif).Assembly.CodeBase).LocalPath;
+			var folder = Path.GetDirectoryName(binPath);
+			var runtime = $"runtimes\\win-{RuntimeInformation.ProcessArchitecture}\\native";
+			var nativePath = Path.Combine(folder, runtime);
+
+			loaded = SetDllDirectory(nativePath);
 			if (!loaded)
 				throw new HeifInteropException("Unable to set native path");
 		}