--- a/src/FileOnQ.Imaging.Heif/Interop/Interop.cs
+++ b/src/FileOnQ.Imaging.Heif/Interop/Interop.cs
@@ -10,19 +10,15 @@ namespace FileOnQ.Imaging.Heif
 		[DllImport("kernel32.dll")]
 		private static extern bool SetDllDirectory(string dllToLoad);
 
-		private static bool loaded;
-
 		internal static void UpdateDllSearchPath()
 		{
-			if (loaded)
-				return;
 
 			var binPath = new Uri(typeof(LibHeif).Assembly.CodeBase).LocalPath;
 			var folder = Path.GetDirectoryName(binPath);
 			var runtime = $"runtimes\\win-{RuntimeInformation.ProcessArchitecture}\\native";
 			var nativePath = Path.Combine(folder, runtime);
 
-			loaded = SetDllDirectory(nativePath);
+			var loaded = SetDllDirectory(nativePath);
 			if (!loaded)
 				throw new HeifInteropException("Unable to set native path");
 		}