--- a/src/FileOnQ.Imaging.Heif/LibHeif/LibHeif.cs
+++ b/src/FileOnQ.Imaging.Heif/LibHeif/LibHeif.cs
@@ -5,6 +5,15 @@ namespace FileOnQ.Imaging.Heif
 {
 	public static unsafe partial class LibHeif
 	{
+#if NET48_OR_GREATER
+		static LibHeif()
+		{
+
+			Interop.UpdateDllSearchPath();
+
+
+		}
+#endif
 		internal static Context* ContextAllocate()
 		{
 			switch (RuntimeInformation.ProcessArchitecture)