… benchmarks/results/primary.net48/FileOnQ.Imaging.Heif.Benchmarks.PrimaryImage-report-full.json
… benchmarks/results/primary.net48/FileOnQ.Imaging.Heif.Benchmarks.PrimaryImage-report.csv
… benchmarks/results/primary.net5.0/FileOnQ.Imaging.Heif.Benchmarks.PrimaryImage-report-full.json
… benchmarks/results/primary.net5.0/FileOnQ.Imaging.Heif.Benchmarks.PrimaryImage-report.csv
… benchmarks/results/primary.net6.0/FileOnQ.Imaging.Heif.Benchmarks.PrimaryImage-report-full.json
… benchmarks/results/primary.net6.0/FileOnQ.Imaging.Heif.Benchmarks.PrimaryImage-report.csv
… benchmarks/results/thumbnail.net48/FileOnQ.Imaging.Heif.Benchmarks.Thumbnail-report-full.json
… benchmarks/results/thumbnail.net48/FileOnQ.Imaging.Heif.Benchmarks.Thumbnail-report.csv
… benchmarks/results/thumbnail.net5.0/FileOnQ.Imaging.Heif.Benchmarks.Thumbnail-report-full.json
… benchmarks/results/thumbnail.net5.0/FileOnQ.Imaging.Heif.Benchmarks.Thumbnail-report.csv
… benchmarks/results/thumbnail.net6.0/FileOnQ.Imaging.Heif.Benchmarks.Thumbnail-report-full.json
… benchmarks/results/thumbnail.net6.0/FileOnQ.Imaging.Heif.Benchmarks.Thumbnail-report.csv