<!-- link your pull request to an open issue -->
Fixes: #65
Fixes: #67 

## Description
The UpdatedllSearchPath method was causing a memory leak. The resolution for this issue was to add a property that is evaluated every time the method is called to determine if the DLL directory has been set. This resolves the memory leak for .NET 4.8 as well as .NET 6.

### Before Benchmarks

|             Method |     Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated | Allocated native memory | Native memory leak |
|------------------- |---------:|---------:|---------:|------:|------:|------:|----------:|------------------------:|-------------------:|
|    Thumbnail_Write | 29.09 ms | 0.211 ms | 0.187 ms |     - |     - |     - |  74,504 B |             5,125,345 B |              234 B |
|  Thumbnail_ToArray | 28.79 ms | 0.171 ms | 0.160 ms |     - |     - |     - |  74,504 B |             5,124,925 B |              234 B |
|   Thumbnail_ToSpan | 29.12 ms | 0.258 ms | 0.215 ms |     - |     - |     - |         - |             5,124,909 B |              234 B |
| Thumbnail_ToStream | 29.11 ms | 0.366 ms | 0.343 ms |     - |     - |     - | 140,816 B |             5,124,925 B |              234 B |


|                Method |    Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 |   Allocated | Allocated native memory | Native memory leak |
|---------------------- |--------:|---------:|---------:|------:|------:|------:|------------:|------------------------:|-------------------:|
|    PrimaryImage_Write | 1.342 s | 0.0043 s | 0.0038 s |     - |     - |     - | 1,951,112 B |           222,030,860 B |              234 B |
|  PrimaryImage_ToArray | 1.343 s | 0.0031 s | 0.0028 s |     - |     - |     - | 1,951,112 B |           222,030,552 B |              234 B |
|   PrimaryImage_ToSpan | 1.342 s | 0.0057 s | 0.0053 s |     - |     - |     - |           - |           222,029,928 B |              234 B |
| PrimaryImage_ToStream | 1.341 s | 0.0043 s | 0.0038 s |     - |     - |     - | 3,894,032 B |           222,030,392 B |              234 B |

### After Benchmarks

|             Method |     Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated | Allocated native memory | Native memory leak |
|------------------- |---------:|---------:|---------:|------:|------:|------:|----------:|------------------------:|-------------------:|
|    Thumbnail_Write | 29.09 ms | 0.211 ms | 0.187 ms |     - |     - |     - |  74,504 B |             5,125,345 B |              - |
|  Thumbnail_ToArray | 28.79 ms | 0.171 ms | 0.160 ms |     - |     - |     - |  74,504 B |             5,124,925 B |              - |
|   Thumbnail_ToSpan | 29.12 ms | 0.258 ms | 0.215 ms |     - |     - |     - |         - |             5,124,909 B |              -  |
| Thumbnail_ToStream | 29.11 ms | 0.366 ms | 0.343 ms |     - |     - |     - | 140,816 B |             5,124,925 B |              - |


|                Method |    Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 |   Allocated | Allocated native memory | Native memory leak |
|---------------------- |--------:|---------:|---------:|------:|------:|------:|------------:|------------------------:|-------------------:|
|    PrimaryImage_Write | 1.342 s | 0.0043 s | 0.0038 s |     - |     - |     - | 1,951,112 B |           222,030,860 B |              - |
|  PrimaryImage_ToArray | 1.343 s | 0.0031 s | 0.0028 s |     - |     - |     - | 1,951,112 B |           222,030,552 B |              - |
|   PrimaryImage_ToSpan | 1.342 s | 0.0057 s | 0.0053 s |     - |     - |     - |           - |           222,029,928 B |              - |
| PrimaryImage_ToStream | 1.341 s | 0.0043 s | 0.0038 s |     - |     - |     - | 3,894,032 B |           222,030,392 B |              - |

## Merge Checklist
- [ ] Added unit or integration tests (if not explain)
- [x] Benchmarks are equivalent or faster