@ahoefling 2022-01-12
@kenny-sellers can you try using a static constructor instead of this technique. You would move all of this code to a static constructor, then we wouldn't need to reference it from the `HeifImage`