@ahoefling 2022-01-12
My local benchmark results show no memory leaks
```
BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19043.1415 (21H1/May2021Update)
AMD Ryzen 9 3950X, 1 CPU, 32 logical and 16 physical cores
  [Host]     : .NET Framework 4.8 (4.8.4420.0), X64 RyuJIT
  Job-GHSIOX : .NET Framework 4.8 (4.8.4420.0), X64 RyuJIT

Runtime=.NET Framework 4.8  InvocationCount=1  LaunchCount=1
UnrollFactor=1

|             Method |     Mean |    Error |   StdDev | Allocated native memory | Native memory leak | Allocated |
|------------------- |---------:|---------:|---------:|------------------------:|-------------------:|----------:|
|    Thumbnail_Write | 33.68 ms | 0.278 ms | 0.260 ms |             5,123,833 B |                  - |  74,504 B |
|  Thumbnail_ToArray | 33.63 ms | 0.322 ms | 0.301 ms |             5,123,277 B |                  - |  74,504 B |
|   Thumbnail_ToSpan | 33.47 ms | 0.261 ms | 0.218 ms |             5,123,533 B |                  - |         - |
| Thumbnail_ToStream | 33.71 ms | 0.576 ms | 0.539 ms |             5,123,277 B |                  - | 140,816 B |
```