@kenny-sellers 2022-01-12
@ahoefling here are benchmarks for all frameworks
## .Net 6 Thumbnail
``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19043.1466 (21H1/May2021Update)
AMD Ryzen 9 5950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK=6.0.101
  [Host]     : .NET 6.0.1 (6.0.121.56705), X64 RyuJIT
  Job-VPNBQD : .NET 6.0.1 (6.0.121.56705), X64 RyuJIT

Runtime=.NET 6.0  InvocationCount=1  LaunchCount=1  
UnrollFactor=1  

```
|             Method |     Mean |    Error |   StdDev | Allocated native memory | Native memory leak | Allocated |
|------------------- |---------:|---------:|---------:|------------------------:|-------------------:|----------:|
|    Thumbnail_Write | 28.72 ms | 0.261 ms | 0.244 ms |             5,124,333 B |                  - |     832 B |
|  Thumbnail_ToArray | 29.86 ms | 0.411 ms | 0.385 ms |             5,123,801 B |                  - |  66,888 B |
|   Thumbnail_ToSpan | 28.75 ms | 0.336 ms | 0.280 ms |             5,123,801 B |                  - |     600 B |
| Thumbnail_ToStream | 28.62 ms | 0.202 ms | 0.158 ms |             5,123,801 B |                  - |  66,952 B |

## .Net 5 Thumbnail
``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19043.1466 (21H1/May2021Update)
AMD Ryzen 9 5950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK=6.0.101
  [Host]     : .NET 5.0.13 (5.0.1321.56516), X64 RyuJIT
  Job-QLUVLW : .NET 5.0.13 (5.0.1321.56516), X64 RyuJIT

Runtime=.NET 5.0  InvocationCount=1  LaunchCount=1  
UnrollFactor=1  

```
|             Method |     Mean |    Error |   StdDev | Allocated native memory | Native memory leak | Allocated |
|------------------- |---------:|---------:|---------:|------------------------:|-------------------:|----------:|
|    Thumbnail_Write | 28.79 ms | 0.237 ms | 0.211 ms |             5,124,365 B |                  - |     288 B |
|  Thumbnail_ToArray | 28.82 ms | 0.276 ms | 0.244 ms |             5,123,813 B |                  - |  66,408 B |
|   Thumbnail_ToSpan | 28.51 ms | 0.297 ms | 0.248 ms |             5,123,817 B |                  - |     120 B |
| Thumbnail_ToStream | 28.90 ms | 0.400 ms | 0.374 ms |             5,123,801 B |                  - |  66,472 B |


## .Net 4.8 Thumbnail
``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19043.1466 (21H1/May2021Update)
AMD Ryzen 9 5950X, 1 CPU, 32 logical and 16 physical cores
  [Host]     : .NET Framework 4.8 (4.8.4420.0), X64 RyuJIT
  Job-GHSIOX : .NET Framework 4.8 (4.8.4420.0), X64 RyuJIT

Runtime=.NET Framework 4.8  InvocationCount=1  LaunchCount=1  
UnrollFactor=1  

```
|             Method |     Mean |    Error |   StdDev | Allocated native memory | Native memory leak | Allocated |
|------------------- |---------:|---------:|---------:|------------------------:|-------------------:|----------:|
|    Thumbnail_Write | 28.93 ms | 0.369 ms | 0.327 ms |             5,123,831 B |                  - |  74,504 B |
|  Thumbnail_ToArray | 28.88 ms | 0.337 ms | 0.316 ms |             5,123,543 B |                  - |  74,504 B |
|   Thumbnail_ToSpan | 28.84 ms | 0.414 ms | 0.387 ms |             5,123,543 B |                  - |         - |
| Thumbnail_ToStream | 29.05 ms | 0.271 ms | 0.254 ms |             5,123,543 B |                  - | 140,816 B |


## .Net 6 Primary
``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19043.1466 (21H1/May2021Update)
AMD Ryzen 9 5950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK=6.0.101
  [Host]     : .NET 6.0.1 (6.0.121.56705), X64 RyuJIT
  Job-RHOQLY : .NET 6.0.1 (6.0.121.56705), X64 RyuJIT

Runtime=.NET 6.0  InvocationCount=1  LaunchCount=1  
UnrollFactor=1  

```
|                Method |    Mean |    Error |   StdDev | Allocated native memory | Native memory leak |   Allocated |
|---------------------- |--------:|---------:|---------:|------------------------:|-------------------:|------------:|
|    PrimaryImage_Write | 1.307 s | 0.0053 s | 0.0050 s |           222,029,704 B |                  - |       848 B |
|  PrimaryImage_ToArray | 1.304 s | 0.0040 s | 0.0035 s |           222,028,932 B |                  - | 1,943,536 B |
|   PrimaryImage_ToSpan | 1.312 s | 0.0047 s | 0.0044 s |           222,029,188 B |                  - |       616 B |
| PrimaryImage_ToStream | 1.315 s | 0.0080 s | 0.0075 s |           222,029,140 B |                  - | 1,943,600 B |

## .Net 5 Primary
``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19043.1466 (21H1/May2021Update)
AMD Ryzen 9 5950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK=6.0.101
  [Host]     : .NET 5.0.13 (5.0.1321.56516), X64 RyuJIT
  Job-TWRSHQ : .NET 5.0.13 (5.0.1321.56516), X64 RyuJIT

Runtime=.NET 5.0  InvocationCount=1  LaunchCount=1  
UnrollFactor=1  

```
|                Method |    Mean |    Error |   StdDev | Allocated native memory | Native memory leak |   Allocated |
|---------------------- |--------:|---------:|---------:|------------------------:|-------------------:|------------:|
|    PrimaryImage_Write | 1.305 s | 0.0021 s | 0.0020 s |           222,029,608 B |                  - |       256 B |
|  PrimaryImage_ToArray | 1.306 s | 0.0044 s | 0.0041 s |           222,029,028 B |                  - | 1,943,008 B |
|   PrimaryImage_ToSpan | 1.306 s | 0.0031 s | 0.0029 s |           222,029,012 B |                  - |        88 B |
| PrimaryImage_ToStream | 1.308 s | 0.0036 s | 0.0034 s |           222,029,172 B |                  - | 1,943,072 B |

## .NET 4.8 Primary
``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19043.1466 (21H1/May2021Update)
AMD Ryzen 9 5950X, 1 CPU, 32 logical and 16 physical cores
  [Host]     : .NET Framework 4.8 (4.8.4420.0), X64 RyuJIT
  Job-GHSIOX : .NET Framework 4.8 (4.8.4420.0), X64 RyuJIT

Runtime=.NET Framework 4.8  InvocationCount=1  LaunchCount=1  
UnrollFactor=1  

```
|                Method |    Mean |    Error |   StdDev | Allocated native memory | Native memory leak |   Allocated |
|---------------------- |--------:|---------:|---------:|------------------------:|-------------------:|------------:|
|    PrimaryImage_Write | 1.312 s | 0.0054 s | 0.0051 s |           222,029,490 B |                  - | 1,951,112 B |
|  PrimaryImage_ToArray | 1.304 s | 0.0028 s | 0.0025 s |           222,029,186 B |                  - | 1,951,112 B |
|   PrimaryImage_ToSpan | 1.306 s | 0.0037 s | 0.0033 s |           222,028,994 B |                  - |           - |
| PrimaryImage_ToStream | 1.310 s | 0.0032 s | 0.0028 s |           222,029,250 B |                  - | 3,894,032 B |